+++ app/scripts/controllers/computed-balances.js
@@ -20,23 +20,34 @@ class ComputedbalancesController {
   }
 
   updateAllBalances () {
-    for (let address in this.accountTracker.store.getState().accounts) {
+    Object.keys(this.balances).forEach((balance) => {
+      const address = balance.address
       this.balances[address].updateBalance()
-    }
+    })
   }
 
   _initBalanceUpdating () {
     const store = this.accountTracker.store.getState()
-    this.addAnyAccountsFromStore(store)
-    this.accountTracker.store.subscribe(this.addAnyAccountsFromStore.bind(this))
+    this.syncAllAccountsFromStore(store)
+    this.accountTracker.store.subscribe(this.syncAllAccountsFromStore.bind(this))
   }
 
-  addAnyAccountsFromStore(store) {
-    const balances = store.accounts
+  syncAllAccountsFromStore(store) {
+    const upstream = Object.keys(store.accounts)
+    const balances = Object.keys(this.balances)
+    .map(address => this.balances[address])
 
+    // Follow new addresses
     for (let address in balances) {
       this.trackAddressIfNotAlready(address)
     }
+
+    // Unfollow old ones
+    balances.forEach(({ address }) => {
+      if (!upstream.includes(address)) {
+        delete this.balances[address]
+      }
+    })
   }
 
   trackAddressIfNotAlready (address) {