+++ app/scripts/lib/account-tracker.js
@@ -38,6 +38,29 @@ class AccountTracker extends EventEmitter {
   // public
   //
 
+  syncWithAddresses (addresses) {
+    const accounts = this.store.getState().accounts
+    const locals = Object.keys(accounts)
+
+    const toAdd = []
+    addresses.forEach((upstream) => {
+      if (!locals.includes(upstream)) {
+        toAdd.push(upstream)
+      }
+    })
+
+    const toRemove = []
+    locals.forEach((local) => {
+      if (!addresses.includes(local)) {
+        toRemove.push(local)
+      }
+    })
+
+    toAdd.forEach(upstream => this.addAccount(upstream))
+    toRemove.forEach(local => this.removeAccount(local))
+    this._updateAccounts()
+  }
+
   addAccount (address) {
     const accounts = this.store.getState().accounts
     accounts[address] = {}